+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 2. Support EIP-1898
 3. Pass through `eth_sendRawTransaction` params
 4. Fix `eth_getTransactionCount`
+5. Add support for `eth_maxPriorityFeePerGas`, `eth_getUncleCountByBlockHash`, `eth_getUncleCountByBlockNumber`
 
 ## v0.5.0